backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
feedparser==6.0.11
google-generativeai==0.7.2
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
aiosqlite==0.20.0
